+++ app/src/main/java/it/niedermann/owncloud/notes/android/fragment/NotePreviewFragment.java
@@ -2,19 +2,12 @@ package it.niedermann.owncloud.notes.android.fragment;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
-import android.support.v4.view.ViewCompat;
-import android.support.v7.widget.SearchView;
-import android.text.SpannableString;
-import android.text.TextUtils;
 import android.text.method.LinkMovementMethod;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.Menu;
-import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
-import android.view.ViewTreeObserver;
-import android.widget.LinearLayout;
 import android.widget.TextView;
 
 import com.yydcdut.rxmarkdown.RxMDTextView;
@@ -24,7 +17,6 @@ import com.yydcdut.rxmarkdown.syntax.text.TextFactory;
 import butterknife.BindView;
 import butterknife.ButterKnife;
 import it.niedermann.owncloud.notes.R;
-import it.niedermann.owncloud.notes.util.DisplayUtils;
 import it.niedermann.owncloud.notes.util.MarkDownUtil;
 import rx.Subscriber;
 import rx.android.schedulers.AndroidSchedulers;
@@ -34,9 +26,6 @@ public class NotePreviewFragment extends BaseNoteFragment {
 
     @BindView(R.id.single_note_content)
     RxMDTextView noteContent;
-
-    private String searchQuery = null;
-    private SearchView searchView;
 
     private static boolean isNewFragment = false;
     public static NotePreviewFragment newInstance(long noteId) {
@@ -53,56 +42,6 @@ public class NotePreviewFragment extends BaseNoteFragment {
         super.onPrepareOptionsMenu(menu);
         menu.findItem(R.id.menu_edit).setVisible(true);
         menu.findItem(R.id.menu_preview).setVisible(false);
-        MenuItem searchMenuItem = menu.findItem(R.id.search);
-
-        searchView = (android.support.v7.widget.SearchView) searchMenuItem.getActionView();
-
-        if (!TextUtils.isEmpty(searchQuery)) {
-            if (!isNewFragment) {
-                searchMenuItem.expandActionView();
-            } else {
-                searchMenuItem.collapseActionView();
-            }
-            searchView.setQuery(searchQuery, true);
-            searchView.clearFocus();
-        } else {
-            searchMenuItem.collapseActionView();
-        }
-
-        final LinearLayout searchEditFrame = searchView.findViewById(android.support.v7.appcompat.R.id
-                .search_edit_frame);
-
-        searchEditFrame.getViewTreeObserver().addOnGlobalLayoutListener(new ViewTreeObserver.OnGlobalLayoutListener() {
-            int oldVisibility = -1;
-            @Override
-            public void onGlobalLayout() {
-                int currentVisibility = searchEditFrame.getVisibility();
-
-                if (currentVisibility != oldVisibility) {
-                    if (currentVisibility != View.VISIBLE) {
-                        colorWithText("");
-                        searchQuery = "";
-                    }
-
-                    oldVisibility = currentVisibility;
-                }
-            }
-
-        });
-
-        searchView.setOnQueryTextListener(new SearchView.OnQueryTextListener() {
-            @Override
-            public boolean onQueryTextSubmit(String query) {
-                return false;
-            }
-
-            @Override
-            public boolean onQueryTextChange(String newText) {
-                searchQuery = newText;
-                colorWithText(newText);
-                return true;
-            }
-        });
     }
 
     @Nullable
@@ -116,9 +55,7 @@ public class NotePreviewFragment extends BaseNoteFragment {
         super.onActivityCreated(savedInstanceState);
         ButterKnife.bind(this, getView());
 
-        if (savedInstanceState != null) {
-            searchQuery = savedInstanceState.getString("searchQuery", "");
-        }
+        setActiveTextView(noteContent);
 
         String content = note.getContent();
         /*
@@ -164,21 +101,4 @@ public class NotePreviewFragment extends BaseNoteFragment {
     protected String getContent() {
         return note.getContent();
     }
-
-    private void colorWithText(String newText) {
-        if (ViewCompat.isAttachedToWindow(noteContent)) {
-            noteContent.setText(DisplayUtils.searchAndColor(noteContent.getText().toString(), new SpannableString
-                            (noteContent.getText()), newText, getResources().getColor(R.color.primary)),
-                    TextView.BufferType.SPANNABLE);
-        }
-    }
-
-    @Override
-    public void onSaveInstanceState(Bundle outState) {
-        super.onSaveInstanceState(outState);
-        if (searchView != null && !TextUtils.isEmpty(searchView.getQuery().toString())) {
-            outState.putString("searchQuery", searchView.getQuery().toString());
-        }
-    }
-
 }
